+++ integration/seccomp.rs
@@ -1,14 +1,14 @@
-#[cfg(target_os = "linux")]
 use std::ffi::CString;
 
-#[cfg(target_os = "linux")]
 use birdcage::{Birdcage, Sandbox};
 
-#[cfg(target_os = "linux")]
-fn main() {
-    // Activate our sandbox.
-    Birdcage::new().lock().unwrap();
+use crate::TestSetup;
 
+pub fn setup() -> TestSetup {
+    TestSetup { sandbox: Birdcage::new(), data: String::new() }
+}
+
+pub fn validate(_data: String) {
     // Ensure `chdir` is allowed.
     let root_path = CString::new("/").unwrap();
     let result = unsafe { libc::chdir(root_path.as_ptr()) };
@@ -24,6 +24,3 @@ fn main() {
     let result = unsafe { libc::syscall(libc::SYS_clone, flags, stack) };
     assert_eq!(result, -1);
 }
-
-#[cfg(not(target_os = "linux"))]
-fn main() {}