+++ src/lib.rs
@@ -7,8 +7,9 @@
 //!
 //! ```rust
 //! use std::fs;
+//! use std::process::Command;
 //!
-//! use birdcage::{Birdcage, Sandbox};
+//! use birdcage::{Birdcage, Exception, Sandbox};
 //! use tempfile::NamedTempFile;
 //!
 //! // Setup our test file.
@@ -17,16 +18,25 @@
 //! // Reads without sandbox work.
 //! fs::read_to_string(file.path()).unwrap();
 //!
+//! // Allow access to our test executable.
+//! let mut sandbox = Birdcage::new();
+//! sandbox.add_exception(Exception::ExecuteAndRead("/bin/cat".into())).unwrap();
+//! let _ = sandbox.add_exception(Exception::ExecuteAndRead("/lib64".into()));
+//! let _ = sandbox.add_exception(Exception::ExecuteAndRead("/lib".into()));
+//!
 //! // Initialize the sandbox; by default everything is prohibited.
-//! Birdcage::new().lock().unwrap();
+//! let mut command = Command::new("/bin/cat");
+//! command.arg(file.path());
+//! let mut child = sandbox.spawn(command).unwrap();
 //!
 //! // Reads with sandbox should fail.
-//! let result = fs::read_to_string(file.path());
-//! assert!(result.is_err());
+//! let status = child.wait().unwrap();
+//! assert!(!status.success());
 //! ```
 
 use std::env;
 use std::path::PathBuf;
+use std::process::{Child, Command};
 
 use crate::error::Result;
 #[cfg(target_os = "linux")]
@@ -64,18 +74,21 @@ pub trait Sandbox: Sized {
     /// symlink's target.
     fn add_exception(&mut self, exception: Exception) -> Result<&mut Self>;
 
-    /// Apply the sandbox restrictions to the current process.
+    /// Setup sandbox and spawn a new process.
+    ///
+    /// This will setup the sandbox in the **CURRENT** process, before launching
+    /// the sandboxee. Since most of the restrictions will also be applied to
+    /// the calling process, it is recommended to create a separate process
+    /// before calling this method. The calling process is **NOT** fully
+    /// sandboxed.
     ///
     /// # Errors
     ///
     /// Sandboxing will fail if the calling process is not single-threaded.
     ///
-    /// Since sandboxing layers are applied in multiple steps, it is possible
-    /// that after a failure some restrictions are still applied. While this
-    /// never allows the process to do things it wasn't capable of doing
-    /// before, it is still recommended to abort the sandboxing process if
-    /// you want to continue operations without a sandbox in place.
-    fn lock(self) -> Result<()>;
+    /// After failure, the calling process might still be affected by partial
+    /// sandboxing restrictions.
+    fn spawn(self, sandboxee: Command) -> Result<Child>;
 }
 
 /// Sandboxing exception rule.